metrics: Move some tests
Those are CLI tests, and should have been there from the start.
