+++ azafea/event_processors/endless/metrics/tests/integration/test_metrics_v2.py
@@ -1818,82 +1818,3 @@ class TestMetrics(IntegrationTest):
 
             assert dbsession.query(InvalidSequence).count() == 0
             assert dbsession.query(UserIsLoggedIn).count() == 0
-
-    def test_parse_old_images(self):
-        from azafea.event_processors.endless.metrics.machine import Machine
-
-        # Create the table
-        self.run_subcommand('initdb')
-        self.ensure_tables(Machine)
-
-        # Insert a machine without parsed image components
-        image_id = 'eos-eos3.7-amd64-amd64.190419-225606.base'
-
-        with self.db as dbsession:
-            dbsession.add(Machine(machine_id='ffffffffffffffffffffffffffffffff', image_id=image_id))
-
-        with self.db as dbsession:
-            machine = dbsession.query(Machine).one()
-            assert machine.image_id == image_id
-            assert machine.image_product is None
-            assert machine.image_branch is None
-            assert machine.image_arch is None
-            assert machine.image_platform is None
-            assert machine.image_timestamp is None
-            assert machine.image_personality is None
-
-        # Parse the image for old machine records
-        self.run_subcommand('test_parse_old_images', 'parse-old-images')
-
-        with self.db as dbsession:
-            machine = dbsession.query(Machine).one()
-            assert machine.image_id == image_id
-            assert machine.image_product == 'eos'
-            assert machine.image_branch == 'eos3.7'
-            assert machine.image_arch == 'amd64'
-            assert machine.image_platform == 'amd64'
-            assert machine.image_timestamp == datetime(2019, 4, 19, 22, 56, 6, tzinfo=timezone.utc)
-            assert machine.image_personality == 'base'
-
-    def test_parse_old_images_skips_already_done(self, capfd):
-        from azafea.event_processors.endless.metrics.machine import Machine
-
-        # Create the table
-        self.run_subcommand('initdb')
-        self.ensure_tables(Machine)
-
-        # Insert a machine without parsed image components
-        image_id = 'eos-eos3.7-amd64-amd64.190419-225606.base'
-
-        with self.db as dbsession:
-            dbsession.add(Machine(machine_id='ffffffffffffffffffffffffffffffff', image_id=image_id,
-                                  image_product='eos', image_branch='eos3.7', image_arch='amd64',
-                                  image_platform='amd64', image_personality='base',
-                                  image_timestamp=datetime(2019, 4, 19, 22, 56, 6,
-                                                           tzinfo=timezone.utc)))
-
-        with self.db as dbsession:
-            machine = dbsession.query(Machine).one()
-            assert machine.image_id == image_id
-            assert machine.image_product == 'eos'
-            assert machine.image_branch == 'eos3.7'
-            assert machine.image_arch == 'amd64'
-            assert machine.image_platform == 'amd64'
-            assert machine.image_timestamp == datetime(2019, 4, 19, 22, 56, 6, tzinfo=timezone.utc)
-            assert machine.image_personality == 'base'
-
-        # Parse the image for old machine records
-        self.run_subcommand('test_parse_old_images_skips_already_done', 'parse-old-images')
-
-        with self.db as dbsession:
-            machine = dbsession.query(Machine).one()
-            assert machine.image_id == image_id
-            assert machine.image_product == 'eos'
-            assert machine.image_branch == 'eos3.7'
-            assert machine.image_arch == 'amd64'
-            assert machine.image_platform == 'amd64'
-            assert machine.image_timestamp == datetime(2019, 4, 19, 22, 56, 6, tzinfo=timezone.utc)
-            assert machine.image_personality == 'base'
-
-        capture = capfd.readouterr()
-        assert 'No machine record with unparsed image ids' in capture.out
